+++ csharp/extractor/Semmle.Extraction.CSharp/Entities/IStatementParentEntity.cs
@@ -1,0 +1,10 @@
+namespace Semmle.Extraction.CSharp.Entities
+{
+    /// <summary>
+    /// Whether this entity is the parent of a top-level statement.
+    /// </summary>
+    public interface IStatementParentEntity : IEntity
+    {
+        bool IsTopLevelParent { get; }
+    }
+}
